alignment move files
