+++ arkouda/pandas/series.py
@@ -11,7 +11,7 @@ from pandas._config import get_option
 from typeguard import typechecked
 
 from arkouda.accessor import CachedAccessor, DatetimeAccessor, StringAccessor
-from arkouda.alignment import lookup
+from arkouda.numpy.alignment import lookup
 from arkouda.numpy.dtypes import bool_scalars, dtype, float64, int64
 from arkouda.numpy.pdarrayclass import RegistrationError, any, argmaxk, create_pdarray, pdarray
 from arkouda.numpy.pdarraycreation import arange, array, full, zeros
