+++ pytest.ini
@@ -26,7 +26,7 @@ testpaths =
     tests/array_api/stats_functions.py
     tests/array_api/utility_functions.py
     tests/array_api/util_functions.py
-    tests/alignment_test.py
+    tests/numpy/alignment_test.py
     tests/array_manipulation_tests.py
     tests/apply_test.py
     tests/bigint_agg_test.py
